# api.yaml
...
swagger: "2.0"
info:
  version: "1.0"
  title: "LatteArt Repository"
basePath: "/api/v1"
paths:
  /test-results:
    get:
      summary: Get test result identifiers.
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              type: object
              properties:
                id:
                  type: string
                  description: Test result ID.
                name:
                  type: string
                  description: Test result name.
    post:
      summary: Create an empty test result.
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            type: object
            properties:
              initialUrl:
                type: string
                description: Initial test target URL.
              name:
                type: string
                description: Test result name.
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              id:
                type: string
                description: Test result ID.
              name:
                type: string
                description: Test result name.
        500:
          description: Internal Server Error
          schema:
            $ref: "#/definitions/ServerError"
  /test-results/{testResultId}:
    get:
      summary: Get a test result.
      parameters:
        - name: testResultId
          in: "path"
          description: Test result ID.
          required: true
          type: string
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/TestResult"
        500:
          description: Internal Server Error
          schema:
            $ref: "#/definitions/ServerError"
    patch:
      summary: Update a test result.
      parameters:
        - name: testResultId
          in: "path"
          description: Test result ID.
          required: true
          type: string
        - in: "body"
          name: "body"
          required: true
          schema:
            type: object
            properties:
              name:
                type: string
                description: Test result name.
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/TestResult"
        500:
          description: Internal Server Error
          schema:
            $ref: "#/definitions/ServerError"
  /test-results/{testResultId}/test-steps:
    post:
      summary: Add a test step.
      parameters:
        - name: "testResultId"
          in: "path"
          description: Test result ID.
          required: true
          type: string
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/TestStepSource"
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              operation:
                $ref: "#/definitions/Operation"
              coverageSource:
                $ref: "#/definitions/CoverageSource"
              inputElementInfo:
                $ref: "#/definitions/InputElementInfo"
        500:
          description: Internal Server Error
          schema:
            $ref: "#/definitions/ServerError"
  /test-results/{testResultId}/test-steps/{sequence}:
    get:
      summary: Get a test step.
      parameters:
        - name: "testResultId"
          in: "path"
          description: Test result ID.
          required: true
          type: string
        - name: "sequence"
          in: "path"
          description: Sequence number of the test step.
          required: true
          type: string
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/TestStepWithSequence"
        500:
          description: Internal Server Error
          schema:
            $ref: "#/definitions/ServerError"
    patch:
      summary: Update links to notes in a test step.
      parameters:
        - name: "testResultId"
          in: "path"
          description: Test result ID.
          required: true
          type: string
        - name: "sequence"
          in: "path"
          description: Sequence number of the test step.
          required: true
          type: string
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/NotesLink"
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/TestStep"
        500:
          description: Internal Server Error
          schema:
            $ref: "#/definitions/ServerError"
  /test-results/{testResultId}/test-steps/{sequence}/compressed-image:
    post:
      summary: Compress screenshot.
      parameters:
        - name: "testResultId"
          in: "path"
          description: Test result ID.
          required: true
          type: string
        - name: "sequence"
          in: "path"
          description: Sequence number of the test step.
          required: true
          type: string
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              imageFileUrl:
                type: string
        500:
          description: Internal Server Error
          schema:
            $ref: "#/definitions/ServerError"
  /test-results/{testResultId}/notes:
    post:
      summary: Add a note.
      parameters:
        - name: "testResultId"
          in: "path"
          description: Test result ID.
          required: true
          type: string
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/NoteSourceWithImageData"
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/Note"
        500:
          description: Internal Server Error
          schema:
            $ref: "#/definitions/ServerError"
  /test-results/{testResultId}/notes/{noteId}:
    get:
      summary: Get a note.
      parameters:
        - name: "testResultId"
          in: "path"
          description: Test result ID.
          required: true
          type: string
        - name: noteId
          in: path
          description: Note ID.
          required: true
          type: string
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/Note"
        500:
          description: Internal Server Error
          schema:
            $ref: "#/definitions/ServerError"
    put:
      summary: Update a note.
      parameters:
        - name: "testResultId"
          in: "path"
          description: Test result ID.
          required: true
          type: string
        - name: noteId
          in: path
          description: Note ID.
          required: true
          type: string
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/NoteSource"
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/Note"
        500:
          description: Internal Server Error
          schema:
            $ref: "#/definitions/ServerError"
    delete:
      summary: Delete a note.
      parameters:
        - name: "testResultId"
          in: "path"
          description: Test result ID.
          required: true
          type: string
        - name: noteId
          in: path
          description: Note ID.
          required: true
          type: string
      responses:
        204:
          description: OK
        500:
          description: Internal Server Error
          schema:
            $ref: "#/definitions/ServerError"
  /test-results/{testResultId}/notes/{noteId}/compressed-image:
    post:
      summary: Compress screenshot.
      parameters:
        - name: "testResultId"
          in: "path"
          description: Test result ID.
          required: true
          type: string
        - name: noteId
          in: path
          description: Note ID.
          required: true
          type: string
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              imageFileUrl:
                type: string
        500:
          description: Internal Server Error
          schema:
            $ref: "#/definitions/ServerError"
  /test-results/{testResultId}/test-scripts:
    get:
      summary: Get test script urls of a test result.
      parameters:
        - name: "testResultId"
          in: "path"
          description: Test result ID.
          required: true
          type: string
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              type: string
              description: Test script url.
    post:
      summary: Save test scripts of a test result.
      parameters:
        - name: "testResultId"
          in: "path"
          description: Test result ID.
          required: true
          type: string
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/TestScript"
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              url:
                type: string
                description: URL of the test scripts.
        500:
          description: Internal Server Error
          schema:
            $ref: "#/definitions/ServerError"
...
definitions:
  ServerError:
    type: object
    properties:
      code:
        type: string
      message:
        type: string
      details:
        type: array
        items:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            target:
              type: string
  TestResult:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
      startTimeStamp:
        type: integer
      endTimeStamp:
        type: integer
      initialUrl:
        type: string
      testSteps:
        type: array
        items:
          $ref: "#/definitions/TestStep"
      coverageSources:
        type: array
        items:
          $ref: "#/definitions/CoverageSource"
      inputElementInfos:
        type: array
        items:
          $ref: "#/definitions/InputElementInfo"
  TestStepSource:
    type: object
    properties:
      input:
        type: string
      type:
        type: string
      elementInfo:
        $ref: "#/definitions/ElementInfo"
      pageSource:
        type: string
      title:
        type: string
      url:
        type: string
      imageData:
        type: string
        format: byte
      windowHandle:
        type: string
      timestamp:
        type: string
  NotesLink:
    type: object
    properties:
      intention:
        type: string
      bugs:
        type: array
        items:
          type: string
      notices:
        type: array
        items:
          type: string
  NoteSource:
    type: object
    properties:
      value:
        type: string
      details:
        type: string
  NoteSourceWithImageData:
    type: object
    properties:
      type:
        type: string
        enum:
          - "intention"
          - "bug"
          - "notice"
      value:
        type: string
      details:
        type: string
      imageData:
        type: string
        format: byte
  TestStepWithSequence:
    type: object
    properties:
      sequence:
        type: number
      operation:
        $ref: "#/definitions/Operation"
      intention:
        type: string
      bugs:
        type: array
        items:
          type: string
      notices:
        type: array
        items:
          type: string
  TestStep:
    type: object
    properties:
      operation:
        $ref: "#/definitions/Operation"
      intention:
        $ref: "#/definitions/Note"
      bugs:
        type: array
        items:
          $ref: "#/definitions/Note"
      notices:
        type: array
        items:
          $ref: "#/definitions/Note"
  Operation:
    type: object
    properties:
      sequence:
        type: number
      input:
        type: string
      type:
        type: string
      elementInfo:
        $ref: "#/definitions/ElementInfo"
      title:
        type: string
      url:
        type: string
      imageFileUrl:
        type: string
      timestamp:
        type: string
      windowHandle:
        type: string
      inputElementInfo:
        type: array
        items:
          $ref: "#/definitions/ElementInfo"
      pageSource:
        type: string
  Note:
    type: object
    properties:
      id:
        type: string
      type:
        type: string
        enum:
          - intention
          - bug
          - notice
      value:
        type: string
      details:
        type: string
      imageFileUrl:
        type: string
      tags:
        type: array
        items:
          type: string
  CoverageSource:
    type: object
    properties:
      title:
        type: string
      url:
        type: string
      screenElements:
        type: array
        items:
          $ref: "#/definitions/ElementInfo"
  InputElementInfo:
    type: object
    properties:
      title:
        type: string
      url:
        type: string
      inputElements:
        type: array
        items:
          $ref: "#/definitions/ElementInfo"
...
  ElementInfo:
    type: object
    properties:
      tagname:
        type: string
      text:
        type: string
      xpath:
        type: string
      value:
        type: string
      checked:
        type: boolean
      attributes:
        type: object
        properties:
          anyAttribute:
            type: string
...
  TestScript:
    type: object
    properties:
      pageObjects:
        type: array
        items:
          type: object
          properties:
            name:
              type: string
            script:
              type: string
      testData:
        type: array
        items:
          type: object
          properties:
            name:
              type: string
            testData:
              type: string
      testSuite:
        type: object
        properties:
          name:
            type: string
          spec:
            type: string
...